+++ src/main/java/ru/gb/buv/spring1L11/SpringDataApplication.java
@@ -1,0 +1,14 @@
+
+package ru.gb.buv.spring1L11;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SpringDataApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SpringDataApplication.class, args);
+	}
+
+}
