+++ src/main/java/ru/gb/buv/spring1L11/config/SecurityConfig.java
@@ -1,0 +1,55 @@
+package ru.gb.buv.spring1L11.config;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.annotation.Bean;
+import org.springframework.http.HttpMethod;
+import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import ru.gb.buv.spring1L11.service.UserService;
+
+@EnableWebSecurity
+@RequiredArgsConstructor
+@Slf4j
+public class SecurityConfig extends WebSecurityConfigurerAdapter {
+    private final UserService userService;
+
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        log.info("Dao Authentication Provider");
+        http.authorizeRequests().antMatchers(HttpMethod.GET, "/**").authenticated()
+                .antMatchers(HttpMethod.POST, "/**").hasAnyRole("ADMIN","SUPERADMIN","MANAGER")
+                .antMatchers(HttpMethod.PUT, "/**").hasAnyRole("ADMIN","SUPERADMIN","MANAGER")
+                .antMatchers(HttpMethod.DELETE, "/**").hasAnyRole("ADMIN","SUPERADMIN","MANAGER")
+                .and()
+                .formLogin();
+
+
+
+
+
+        /*http.authorizeRequests()
+                .antMatchers("/**").authenticated()
+                .antMatchers("/user_info").authenticated()
+                .antMatchers("/admin/**").hasAnyRole("ADMIN", "SUPERADMIN") // ROLE_ADMIN, ROLE_SUPERADMIN
+                .anyRequest().permitAll()
+                .and()
+                .formLogin();*/
+    }
+
+    @Bean
+    public BCryptPasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+
+    @Bean
+    public DaoAuthenticationProvider daoAuthenticationProvider() {
+        DaoAuthenticationProvider authenticationProvider = new DaoAuthenticationProvider();
+        authenticationProvider.setPasswordEncoder(passwordEncoder());
+        authenticationProvider.setUserDetailsService(userService);
+        return authenticationProvider;
+    }
+}
