+++ src/main/java/ru/gb/buv/spring1L11/dto/ProductDto.java
@@ -1,0 +1,14 @@
+package ru.gb.buv.spring1L11.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+    //Правильное ДТО
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ProductDto {
+    private Long id;
+    private String title;
+    private Long cost;
+}
