+++ src/main/java/ru/gb/buv/spring1L11/entity/Product.java
@@ -1,0 +1,24 @@
+
+package ru.gb.buv.spring1L11.entity;
+
+import lombok.*;
+import javax.persistence.*;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+@Table(name = "products")
+public class Product {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)//-последовательно генерит значения
+    @Column(name = "id")
+    private Long id;
+
+    @Column(name = "title")
+    private String title;
+
+    @Column(name = "cost")
+    private Long cost;
+}
