+++ src/main/java/ru/gb/buv/spring1L11/entity/User.java
@@ -1,0 +1,31 @@
+package ru.gb.buv.spring1L11.entity;
+
+import lombok.Data;
+
+import javax.persistence.*;
+import java.util.Collection;
+
+@Entity
+@Data
+@Table(name = "users")
+public class User {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    private Long id;
+
+    @Column(name = "username")
+    private String username;
+
+    @Column(name = "password")
+    private String password;
+
+
+
+
+    @ManyToMany
+    @JoinTable(name = "users_roles",
+            joinColumns = @JoinColumn(name = "user_id"),
+            inverseJoinColumns = @JoinColumn(name = "role_id"))
+    private Collection<Roles> roles;
+}
