+++ src/main/java/ru/gb/buv/spring1L11/exceptions/ResourceNotFoundException.java
@@ -1,0 +1,7 @@
+package ru.gb.buv.spring1L11.exceptions;
+
+public class ResourceNotFoundException extends RuntimeException{
+    public ResourceNotFoundException(String message){
+        super(message);
+    }
+}
