+++ src/main/java/ru/gb/buv/spring1L11/exceptions/ValidationException.java
@@ -1,0 +1,15 @@
+package ru.gb.buv.spring1L11.exceptions;
+
+import lombok.Data;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Data
+public class ValidationException extends RuntimeException{
+    private List<String> errorFieldMessages;
+    public ValidationException(List<String> errorFieldMessages){
+        super(errorFieldMessages.stream().collect(Collectors.joining(",  ")));
+        this.errorFieldMessages = errorFieldMessages;
+    }
+}
