--- src/main/java/ru/gb/buv/spring_lesson10/SpringDataApplication.java
@@ -1,14 +1,0 @@
-
-package ru.gb.buv.spring_lesson10;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SpringDataApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SpringDataApplication.class, args);
-	}
-
-}
