--- src/main/java/ru/gb/buv/spring_lesson10/exceptions/FieldsValidatorError.java
@@ -1,16 +1,0 @@
-package ru.gb.buv.spring_lesson10.exceptions;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@NoArgsConstructor
-@Data
-public class FieldsValidatorError {
-    private List<String> errorFieldMessages;
-
-    public FieldsValidatorError(List<String> errorFieldMessages){
-        this.errorFieldMessages = errorFieldMessages;
-    }
-}
