--- src/main/java/ru/gb/buv/spring_lesson10/exceptions/ResourceNotFoundException.java
@@ -1,7 +1,0 @@
-package ru.gb.buv.spring_lesson10.exceptions;
-
-public class ResourceNotFoundException extends RuntimeException{
-    public ResourceNotFoundException(String message){
-        super(message);
-    }
-}
