--- src/main/java/ru/gb/buv/spring_lesson10/exceptions/ValidationException.java
@@ -1,15 +1,0 @@
-package ru.gb.buv.spring_lesson10.exceptions;
-
-import lombok.Data;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-@Data
-public class ValidationException extends RuntimeException{
-    private List<String> errorFieldMessages;
-    public ValidationException(List<String> errorFieldMessages){
-        super(errorFieldMessages.stream().collect(Collectors.joining(",  ")));
-        this.errorFieldMessages = errorFieldMessages;
-    }
-}
