+++ src/main/resources/application.yml
@@ -9,10 +9,6 @@ spring:
     driver-class-name: org.h2.Driver
     username: sa
     password:
-      #//проблемы с показом одного продукта:(
-  #spring:
-    #jackson:
-      #serialization: fail-on-empty-beans= false
   jpa:
     #Печатает в лог-консоль произведенные запросы
     show-sql: true
@@ -21,12 +17,11 @@ spring:
     #Авто-создание БД
     hibernate:
       ddl-auto: create
-
   #Что бы могли подключиться к базе H2 через браузер
-  #localhost:8080/h2-console
+  #localhost:8080/myapp/h2-console
   h2:
     console:
       enabled: true
   #
-  liquibase:
-    change-log: classpath:/db/changelog/db.changelog-master.xml
+  #liquibase:
+   # change-log: classpath:/db/changelog/db.changelog-master.xml
