+++ src/test/java/ru/gb/buv/spring1L11/SpringDataApplicationTests.java
@@ -1,0 +1,13 @@
+package ru.gb.buv.spring1L11;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SpringDataApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
