--- src/test/java/ru/gb/buv/spring_lesson10/SpringDataApplicationTests.java
@@ -1,13 +1,0 @@
-package ru.gb.buv.spring_lesson10;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SpringDataApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
