+++ задание
@@ -1,0 +1,17 @@
+Урок 11. Spring Security
+Задание
+1. Создать страницу со списком товаров, на которой
+можно добавлять позиции и редактировать существующие.
+На эту страницу должны иметь доступ админы и менеджеры.
+
+2. Создать страницу со списком всех пользователей,
+к которой имеют доступ только админы.
+
+3. * Добавить роль суперадмина и дать ему возможность
+создавать новых пользователей и указывать роли существующим.
+
+
+USER
+MANGER
+ADMIN
+SUPERADMIN
